requirements.txt
pillow>=9.0.0
requests>=2.28.0
openai>=1.0.0
pydantic>=2.0.0
tqdm>=4.65.0
